// backend/src/modules/auth/auth.schema.ts
import { z } from "zod";

export const registerSchema = z.object({
  name: z.string().min(2, "Name must be at least 2 characters"),
  email: z.string().email("Invalid email address"),
  password: z.string().min(6, "Password must be at least 6 characters"),
});
export const loginSchema = z.object({
  email: z.string().email(),
  password: z.string().min(6),
});
export type RegisterInput = z.infer<typeof registerSchema>;
export type loginSchema = z.infer<typeof loginSchema>;